fix: ProfileCard 중복 컴포넌트 이름 수정
