+++ src/components/profile/ProfileFormLayout/styles.module.scss
@@ -1,6 +1,6 @@
 @import '@/styles/variables';
 
-.profileCard {
+.profileLayout {
   width: 90%;
   height: 400px;
   border-radius: 40px;
