+++ src/components/profile/index.ts
@@ -1,3 +1,3 @@
-export { ProfileCard } from './ProfileCard/ProfileCard'
+export { ProfileFormLayout } from './ProfileFormLayout/ProfileFormLayout'
 export { ProfileSlider } from './ProfileSlider/ProfileSlider'
 export { SliderButtons } from './SliderButtons/SliderButtons'
